changed styling, added effect to slide in details.
