+++ src/components/RestaurantList.css
@@ -1,49 +1,3 @@
 section {
   background-color: #efefef;
 }
-article {
-  background-color: #fff;
-  margin: 10px 0 10px 0;
-  text-align: left;
-  color: grey;
-  padding: 15px;
-}
-
-article img {
-  width: 100%;
-  height: 150px;
-  object-fit: cover;
-  overflow: hidden;
-}
-
-article h3 {
-  color: #333;
-  margin: 5px 0 5px 0;
-}
-
-.dollarSigns {
-  display: inline;
-  margin-right: 5px;
-}
-
-.top-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-}
-
-.tag {
-  background-color: #efefef;
-  display: flex;
-  align-items: center;
-  padding: 0 8px 0 8px;
-  height: 28px;
-}
-
-.star {
-  color: black;
-  size: 18px;
-  margin: 0 4px 0 4px;
-  align-items: ;
-}
